' < sabrown |X| D11_11_29_19_43_52 | Remote fiona > Add scgnu.h and refactor accordingly'

+++ source/score/scfnct.c
@@ -360,7 +360,7 @@ static void _SC_timeout_error(int sig)
 
     char s[MAXLINE];
 
-    snprintf(s, MAXLINE, "\nProcess %d timed out\n\n", getpid());
+    snprintf(s, MAXLINE, "\nProcess %d timed out\n\n", (int) getpid());
 
 /* NOTE: do not use io_puts here because timeouts often occur because
  * of memory deadlocks and io_puts allocates more memory so we would
